Add tests for is_spd. * test_fast_maths.py has been renamed to test_extmath.py * Only use relative tolerance in is_spd to check that the matrix is symmetric
